pyproject.toml
[project]
name = "ai-news-integrator"
version = "1.0.0"
description = "AI Domain News Integrator - Intelligent aggregation tool powered by LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "ddgs>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv",
    "PyYAML",
    "feedparser>=6.0.0",
    "tenacity>=8.0.0",
    "jinja2>=3.0.0",
]

[tool.setuptools]
packages = ["ai_news"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
